+++ src/contexts/WalletContext/useKeplr/useKeplr.ts
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { Connection, ProviderInfo } from "../types";
 import { BaseChain } from "types/aws";
-import { Dwindow } from "types/ethereum";
+import { Dwindow } from "types/window";
 import {
   UnsupportedChainError,
   WalletError,
