+++ src/helpers/keplr.ts
@@ -3,7 +3,7 @@ import { fromHex, toBase64 } from "@cosmjs/encoding";
 import { KeplrWalletConnectV1 } from "@keplr-wallet/wc-client";
 import WalletConnect from "@walletconnect/client/";
 import { ProviderId } from "contexts/WalletContext/types";
-import { Dwindow } from "types/ethereum";
+import { Dwindow } from "types/window";
 import { JUNO_LCD } from "constants/env";
 import { WC_BRIDGE } from "constants/urls";
