+++ src/types/window.ts
@@ -1,0 +1,11 @@
+import { Keplr } from "@keplr-wallet/types";
+
+export interface Dwindow extends Window {
+  xfi?: {
+    ethereum?: { isMetaMask?: boolean; [index: string]: any };
+    terra?: any;
+  };
+  ethereum?: any;
+  BinanceChain?: any;
+  keplr?: Keplr;
+}
